Run wasp setup conditionally

+++ sandbox/bootstrap.sh
@@ -51,12 +51,20 @@ if [[ "$OSTYPE" != "darwin"* ]]; then
   chown -R 65532:65532 data
 fi
 
+source .env
+echo ${COMPOSE_PROFILES}
+
+# Check if compose profiles contains wasp
+if [[ ${COMPOSE_PROFILES} != *"wasp"* ]]; then
+  echo "Setup done, run docker compose up -d to start the network."
+  exit
+fi
+
 # Bootstrap chain
 echo "Bootstrap chain..."
 docker compose --profile wasp --profile bootstrap-chain up -d && docker compose logs bootstrap-chain -f && docker compose logs fund-chain-account -f
 echo "Configuring toolkit..."
 chain_address=$(jq -r '.chains."glass-chain"' data/sandboxdb/wasp-cli/config.json)
-source .env
 sed -e "s/<httpPort>/${HTTP_PORT:-80}/g" -e "s/<chainAddress>/$chain_address/g" assets/evm-toolkit/networks.json > data/sandboxdb/evm-toolkit/networks.json
 echo "Bootstraping and funding chain done. Cleaning up..."
 docker compose --profile wasp --profile bootstrap-chain down
